Move account stuff from `UserService` to `carp.common` `AccountService`.
