+++ carp.common/src/commonMain/kotlin/dk/cachet/carp/common/users/NotifyUserService.kt
@@ -1,4 +1,4 @@
-package dk.cachet.carp.deployment.domain
+package dk.cachet.carp.common.users
 
 import dk.cachet.carp.common.EmailAddress
 import dk.cachet.carp.common.UUID
