+++ carp.deployment.core/src/commonMain/kotlin/dk/cachet/carp/deployment/infrastructure/UserServiceRequest.kt
@@ -1,13 +1,9 @@
 package dk.cachet.carp.deployment.infrastructure
 
-import dk.cachet.carp.common.EmailAddress
 import dk.cachet.carp.common.UUID
 import dk.cachet.carp.common.ddd.createServiceInvoker
-import dk.cachet.carp.common.ddd.createServiceInvokerOverloaded
 import dk.cachet.carp.common.ddd.ServiceInvoker
-import dk.cachet.carp.common.users.Account
 import dk.cachet.carp.common.users.AccountIdentity
-import dk.cachet.carp.common.users.Username
 import dk.cachet.carp.deployment.application.UserService
 import dk.cachet.carp.deployment.domain.users.Participation
 import kotlinx.serialization.Serializable
@@ -19,22 +15,6 @@ import kotlinx.serialization.Serializable
 @Serializable
 sealed class UserServiceRequest
 {
-    @Serializable
-    @Suppress( "RemoveExplicitTypeArguments" ) // Needed for JS build to work.
-    data class CreateAccountWithUsername( val username: Username ) :
-        UserServiceRequest(),
-        ServiceInvoker<UserService, Account>
-            by createServiceInvokerOverloaded<UserService, @ParameterName( "username" ) Username, Account>(
-                UserService::createAccount, "username", username )
-
-    @Serializable
-    @Suppress( "RemoveExplicitTypeArguments" ) // Needed for JS build to work.
-    data class CreateAccountWithEmailAddress( val emailAddress: EmailAddress ) :
-        UserServiceRequest(),
-        ServiceInvoker<UserService, Unit>
-            by createServiceInvokerOverloaded<UserService, @ParameterName( "emailAddress" ) EmailAddress, Unit>(
-                UserService::createAccount, "emailAddress", emailAddress )
-
     @Serializable
     data class AddParticipation( val studyDeploymentId: UUID, val identity: AccountIdentity ) :
         UserServiceRequest(),
