+++ carp.deployment.core/src/commonTest/kotlin/dk/cachet/carp/deployment/infrastructure/UserServiceRequestsTest.kt
@@ -1,9 +1,7 @@
 package dk.cachet.carp.deployment.infrastructure
 
-import dk.cachet.carp.common.EmailAddress
 import dk.cachet.carp.common.UUID
 import dk.cachet.carp.common.ddd.ServiceInvoker
-import dk.cachet.carp.common.users.Username
 import dk.cachet.carp.common.users.UsernameAccountIdentity
 import dk.cachet.carp.deployment.application.UserService
 import dk.cachet.carp.deployment.application.UserServiceMock
@@ -19,8 +17,6 @@ class UserServiceRequestsTest
     companion object
     {
         val requests: List<UserServiceRequest> = listOf(
-            UserServiceRequest.CreateAccountWithUsername( Username( "Test" ) ),
-            UserServiceRequest.CreateAccountWithEmailAddress( EmailAddress( "test@test.com" ) ),
             UserServiceRequest.AddParticipation( UUID.randomUUID(), UsernameAccountIdentity( "Test" ) ),
             UserServiceRequest.GetParticipationsForStudyDeployment( UUID.randomUUID() )
         )
